+++ index.js
@@ -11,9 +11,16 @@ const requesInfoRaw = await queryMySQL(sqlGetRequestInfo);
 
 let requesInfo = {};
 requesInfoRaw.forEach((item) => {
-  const { name_id, weibo_uid, weibo_containerid, bili_channel_id, douyin_cid } =
-    item;
+  const {
+    name_id,
+    name,
+    weibo_uid,
+    weibo_containerid,
+    bili_channel_id,
+    douyin_cid,
+  } = item;
   requesInfo[name_id] = {
+    name: name,
     weibo_uid: weibo_uid,
     weibo_containerid: weibo_containerid,
     bili_channel_id: bili_channel_id,
@@ -25,6 +32,7 @@ for (let item in requesInfo) {
   const sqlInsertData = 'INSERT INTO web_data SET ?';
   const insertData = await getAllData(
     parseInt(item),
+    requesInfo[item]['name'],
     requesInfo[item]['weibo_uid'],
     requesInfo[item]['weibo_containerid'],
     requesInfo[item]['bili_channel_id'],